lambdas/exif/pyproject.toml
[project]
name = "exif-lambda"
version = "0.1.0"
description = "Lambda function to extract EXIF data from images"
requires-python = ">=3.12"
dependencies = [
    # Add your dependencies here
    # Example: "pillow>=10.0.0"
    "boto3>=1.40.56",
    "pillow>=12.0.0",
]
